Single-time modified treatment policy

MTP represents a deterministic single-time modified treatment policy.

It supports two common cases:

discrete treatments, where the post-policy mass function can be computed by summing over pre-images;

continuous treatments, where the policy is assumed piecewise invertible and the post-policy density can be computed via a change-of-variables formula.

For continuous treatments, users supply a collection of policy regions and, for each region, the forward map, inverse map, and inverse derivative. This is the setup used for piecewise smooth LMTPs in Díaz et al. (2023).

Public fields

treatment_type

Either "continuous" or "discrete".

name

Optional name for the policy.

region_predicates

List of functions (A, H) -> logical defining the regions of the observed treatment space.

policy_pieces

List of functions (A, H) -> A_star defining the forward map on each region.

inverse_map_pieces

List of functions (A_star, H) -> A giving the inverse map on each region.

inverse_deriv_pieces

List of functions (A_star, H) -> numeric giving the derivative of the inverse map on each region.

support

Optional vector of support values for discrete treatment.

Methods

Public methods

MTP$new()

MTP$apply_policy()

MTP$ratio_from_density()

MTP$gd_from_density()

MTP$print()

MTP$clone()

Method new()

Create a new single-time MTP.

Usage

MTP$new(
  treatment_type = c("continuous", "discrete"),
  region_predicates,
  policy_pieces,
  inverse_map_pieces = NULL,
  inverse_deriv_pieces = NULL,
  support = NULL,
  name = NULL
)

Arguments

treatment_type

Either "continuous" or "discrete".

region_predicates

List of region-membership functions.

policy_pieces

List of forward-map functions.

inverse_map_pieces

List of inverse-map functions. Required for continuous treatments.

inverse_deriv_pieces

List of inverse-derivative functions. Required for continuous treatments.

support

Optional support for discrete treatments.

name

Optional name.

Method apply_policy()

Apply the policy to observed treatment values.

Usage

MTP$apply_policy(A_vec, H_df)

Arguments

A_vec

Vector of observed treatment values.

H_df

Data frame of histories H_t.

Returns

Vector of shifted treatment values.

Method ratio_from_density()

Evaluate the density ratio r(a, h) = g^d(a | h) / g(a | h) from an observed-data density function.

Usage

MTP$ratio_from_density(A_vec, H_df, density_fun)

Arguments

A_vec

Vector of observed treatment values.

H_df

Data frame of histories H_t.

density_fun

Function (A_vec, H_df) -> g(a | h).

Returns

Numeric vector of density ratios.

Method gd_from_density()

Compute the post-policy density or mass function g^d(a | h) from an observed-data density or mass function.

Usage

MTP$gd_from_density(A_vec, H_df, density_fun)

Arguments

A_vec

Vector of treatment values at which to evaluate the post-policy distribution.

H_df

Data frame of histories H_t.

density_fun

Function (A_vec, H_df) -> g(a | h).

Returns

Numeric vector.

Method print()

Print a compact policy summary.

Usage

MTP$print(...)

Returns

The object invisibly.

Method clone()

The objects of this class are cloneable with this method.

Usage

MTP$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## -------------------------------------------------
## Continuous MTP: simple additive shift A -> A + 1
## -------------------------------------------------

n <- 100
H <- data.frame(W = rnorm(n))
A <- rnorm(n)

mtp <- MTP$new(
  treatment_type = "continuous",
  region_predicates = list(
    function(A, H) rep(TRUE, length(A))
  ),
  policy_pieces = list(
    function(A, H) A + 1
  ),
  inverse_map_pieces = list(
    function(A_star, H) A_star - 1
  ),
  inverse_deriv_pieces = list(
    function(A_star, H) rep(1, length(A_star))
  ),
  name = "shift_plus_one"
)

mtp
#> MTP
#>   name: shift_plus_one
#>   treatment_type: continuous
#>   n_regions: 1

# Apply policy
A_star <- mtp$apply_policy(A, H)
head(cbind(A, A_star))
#>                A      A_star
#> [1,] -2.32640953 -1.32640953
#> [2,]  2.19299800  3.19299800
#> [3,] -1.08247998 -0.08247998
#> [4,] -0.50636097  0.49363903
#> [5,]  0.09591896  1.09591896
#> [6,] -0.12016526  0.87983474

## -------------------------------------------------
## Discrete MTP: flip binary treatment
## -------------------------------------------------

A_bin <- rbinom(n, 1, 0.5)

mtp_flip <- MTP$new(
  treatment_type = "discrete",
  region_predicates = list(
    function(A, H) rep(TRUE, length(A))
  ),
  policy_pieces = list(
    function(A, H) 1 - A
  ),
  support = c(0, 1),
  name = "flip_binary"
)

mtp_flip
#> MTP
#>   name: flip_binary
#>   treatment_type: discrete
#>   n_regions: 1
#>   support: 0, 1

A_star <- mtp_flip$apply_policy(A_bin, H)
head(cbind(A_bin, A_star))
#>      A_bin A_star
#> [1,]     0      1
#> [2,]     1      0
#> [3,]     1      0
#> [4,]     0      1
#> [5,]     0      1
#> [6,]     0      1

## -------------------------------------------------
## Discrete MTP 2: decrement non-negative integer treatment
## -------------------------------------------------

A_pois <- rpois(n, 2)

mtp_decrement <- MTP$new(
  treatment_type = "discrete",
  region_predicates = list(
    function(A, H) A > 0,
    function(A, H) A == 0
  ),
  policy_pieces = list(
    function(A, H) A - 1,
    function(A, H) A
  ),
  name = "decrement_integer"
)
#> Warning: Discrete `MTP` created without `support`. `gd_from_density()` will require `support`.

mtp_decrement
#> MTP
#>   name: decrement_integer
#>   treatment_type: discrete
#>   n_regions: 2

A_star <- mtp_decrement$apply_policy(A_pois, H)
head(cbind(A_pois, A_star))
#>      A_pois A_star
#> [1,]      3      2
#> [2,]      2      1
#> [3,]      3      2
#> [4,]      0      0
#> [5,]      0      0
#> [6,]      4      3